feat: заводит папку blocks для компонентов
