+++ src/components/blocks/place-card/place-card.tsx
@@ -1,27 +1,28 @@
 import {Link} from 'react-router-dom';
-import {ListOfferProps} from '../../types/common-types';
+import {ListOfferProps} from '../../../types/common-types';
 
-type FavoriteCardProps = Pick<ListOfferProps, 'id' | 'isPremium' | 'previewImage' | 'price' | 'title' | 'rating' | 'type'>;
-export function FavoriteCard(props: FavoriteCardProps) {
+type PlaceCardProps = Pick<ListOfferProps, 'id' | 'isPremium' | 'previewImage' | 'price' | 'title' | 'rating' | 'type'>;
+
+function PlaceCard(props: PlaceCardProps) {
   return (
-    <article className="favorites__card place-card">
+    <article className="cities__card place-card">
       {props.isPremium && <div className="place-card__mark"><span>Premium</span></div>}
-      <div className="favorites__image-wrapper place-card__image-wrapper">
+      <div className="cities__image-wrapper place-card__image-wrapper">
         <Link to={`/offer/${props.id}`}>
-          <img className="place-card__image" src="img/apartment-small-03.jpg" width="150" height="110" alt="Place image"/>
+          <img className="place-card__image" src={props.previewImage} width="260" height="200" alt="Place image"/>
         </Link>
       </div>
-      <div className="favorites__card-info place-card__info">
+      <div className="place-card__info">
         <div className="place-card__price-wrapper">
           <div className="place-card__price">
             <b className="place-card__price-value">&euro;{props.price}</b>
             <span className="place-card__price-text">&#47;&nbsp;night</span>
           </div>
-          <button className="place-card__bookmark-button place-card__bookmark-button--active button" type="button">
+          <button className="place-card__bookmark-button button" type="button">
             <svg className="place-card__bookmark-icon" width="18" height="19">
               <use xlinkHref="#icon-bookmark"></use>
             </svg>
-            <span className="visually-hidden">In bookmarks</span>
+            <span className="visually-hidden">To bookmarks</span>
           </button>
         </div>
         <div className="place-card__rating rating">
@@ -38,3 +39,5 @@ export function FavoriteCard(props: FavoriteCardProps) {
     </article>
   );
 }
+
+export {PlaceCard};
